A huge update, which sets the ground for awesome future functionality.
BLogger now has the following structure:
A base class, which handles text formatting, as well as basic
functions e.g 'log()'.
An implementation class, which inherits the base and is responsible
for the actual logging.
This design allows for easy and natural modularity. For example, now
we can have multiple versions of the logger - a simple single-threaded
implementation, as well as an advanced async implementation, which
utilizes a dedicated thread-pool with a message queue.
This is especially important as IO operations are extremely expensive
for any real-time application.