An early prototype of async version of BLogger is finally here!
It is more of an alpha version, however it's already usable!
Some facts:
It's about 50 times faster than the blocking version of BLogger,
taking about 2 microseconds per log message with args(!!).
The only downside is that it takes quite a bit of memory, my benchmark
program was taking about 10 megabytes of RAM at peaks & even more in debug.