A small bug fix.
There was a bug where if you hadn't set a pattern/had an empty pattern
the logger would still try to log using that pattern and would end up
logging '\n' and nothing else. This should obviously not be the case
and the expected behavior is to ignore all the messages if the pattern
is empty.